Move the uid_catalog path helpers to the upgrade utils
